Closes #2930.

---
PHP 7.1.4
Apache+nginx
mySQL 5.6.34

1) Basic Test Form
    * `develop` 854ms
    * `issue#2930` 700ms
1) Event Registration Form
    * `develop` 912ms
    * `issue#2930` 594ms
1) CL, L&S, MP
    * `develop` 2.94s
    * `issue#2930` 1.05s

---
PHP 5.2.4
Apache
mySQL 5.6.34

1) Basic Test Form
    * `develop` 2.67s
    * `issue#2930`  2.38s
1) Event Registration Form
    * `develop` 3.81s
    * `issue#2930` 2.15s
1) CL, L&S, MP
    * `develop` 4.37s
    * `issue#2930` 2.24s